refactor(mpi): agrega os en get version 2
