+++ core-v2/mpi/paciente/paciente.controller.ts
@@ -7,6 +7,9 @@ import { isSelected } from '@andes/core';
 import * as config from '../../../config';
 import { getObraSocial } from '../../../modules/obraSocial/controller/obraSocial';
 import { PacienteCtr } from './paciente.routes';
+import { geoReferenciar, getBarrio } from '@andes/georeference';
+import * as Barrio from '../../../core/tm/schemas/barrio';
+import * as configPrivate from '../../../config.private';
 const ObjectId = Types.ObjectId;
 /**
  * Crea un objeto paciente
@@ -153,7 +156,7 @@ export function isMatchingAlto(sugeridos: any[]) {
 export async function multimatch(searchText: string, filter: any, options?: any) {
     const ExpRegFilter = /([-_()\[\]{}+?*.$\^|¨`´~,:#<>¡!\\])/g;
     let words: any = searchText.replace(ExpRegFilter, '');
-    words = searchText.trim().toLowerCase().split(' ');
+    words = words.trim().toLowerCase().split(' ');
     let andQuery = [];
     words.forEach(w => {
         andQuery.push({ tokens: RegExp(`^${w}`) });
@@ -234,3 +237,40 @@ export async function linkPacientesDuplicados(req, paciente) {
     }
 }
 
+
+/**
+ * * Segun la entrada, retorna un Point con las coordenadas de geo referencia o null.
+ * @param data debe contener direccion y localidad.
+ */
+
+export const updateGeoreferencia = async (paciente: IPacienteDoc) => {
+    try {
+        let direccion: any = paciente.direccion;
+        // (valores de direccion fueron modificados): están completos?
+        if (direccion[0].valor && direccion[0].ubicacion.localidad && direccion[0].ubicacion.provincia) {
+            let dir = direccion[0].valor + ', ' + direccion[0].ubicacion.localidad.nombre + ', ' + direccion[0].ubicacion.provincia.nombre;
+            const geoRef: any = await geoReferenciar(dir, configPrivate.geoKey);
+            // georeferencia exitosa?
+            if (geoRef && Object.keys(geoRef).length) {
+                direccion[0].geoReferencia = [geoRef.lat, geoRef.lng];
+                let nombreBarrio = await getBarrio(geoRef, configPrivate.geoNode.host, configPrivate.geoNode.auth.user, configPrivate.geoNode.auth.password);
+                // consulta exitosa?
+                if (nombreBarrio) {
+                    const barrioPaciente = await Barrio.findOne().where('nombre').equals(RegExp('^.*' + nombreBarrio + '.*$', 'i'));
+                    if (barrioPaciente) {
+                        direccion[0].ubicacion.barrio = barrioPaciente;
+                    }
+                }
+            } else {
+                direccion[0].geoReferencia = null;
+                direccion[0].ubicacion.barrio = null;
+            }
+        }
+        if (direccion[0].georeferencia) {
+            paciente = set(paciente, direccion);
+            PacienteCtr.update(paciente.id, paciente, configPrivate.userScheduler as any);
+        }
+    } catch (err) {
+        return (err);
+    }
+};
